Refactor trade tracking
