+++ src/ui/signal_entry.h
@@ -1,0 +1,10 @@
+#pragma once
+
+struct SignalEntry {
+    double time;
+    double price;
+    double value1;
+    double value2;
+    int type;
+};
+
