+++ src/ui/signals_window.h
@@ -6,14 +6,7 @@
 
 #include "core/candle.h"
 #include "app.h"
-
-struct SignalEntry {
-    double time;
-    double price;
-    double value1;
-    double value2;
-    int type;
-};
+#include "ui/signal_entry.h"
 
 void DrawSignalsWindow(
     std::string& strategy,
@@ -23,10 +16,7 @@ void DrawSignalsWindow(
     double& overbought,
     bool& show_on_chart,
     std::vector<SignalEntry>& signal_entries,
-    std::vector<double>& buy_times,
-    std::vector<double>& buy_prices,
-    std::vector<double>& sell_times,
-    std::vector<double>& sell_prices,
+    std::vector<App::AppContext::TradeEvent>& trades,
     const std::map<std::string, std::map<std::string, std::vector<Core::Candle>>>& all_candles,
     const std::string& active_pair,
     const std::string& selected_interval,
